backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anouar Aissaoui <anouar.aissaoui@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
django = "^4.2.4"
environs = {extras = ["django"], version = "^9.5.0"}
djangorestframework = "^3.14.0"
json-log-formatter = "^0.5.2"
pycountry = "^22.3.5"
django-rest-knox = "^4.2.0"
django-cors-headers = "^4.3.0"
psycopg = {extras = ["binary"], version = "^3.1.12"}
numpy = "^1.26.1"
pandas = "^2.1.1"
django-storages = {extras = ["s3"], version = "^1.14.2"}
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
bandit = "^1.7.5"
pytest-django = "^4.5.2"
pytest = "^7.4.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings.settings"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
